test(analytics): make isolated-stack browser validation explicit
